#41 refactor: naming 수정 및 파일 위치 수정
