+++ src/test/java/org/com/itpple/spot/server/service/PotServiceTest.java
@@ -11,8 +11,8 @@ import java.util.List;
 import org.com.itpple.spot.server.constant.PotType;
 import org.com.itpple.spot.server.constant.Role;
 import org.com.itpple.spot.server.dto.Location;
-import org.com.itpple.spot.server.dto.SearchCondition.CircleSearchRange;
-import org.com.itpple.spot.server.dto.SearchCondition.SearchRange;
+import org.com.itpple.spot.server.dto.pot.SearchCondition.CircleSearchRange;
+import org.com.itpple.spot.server.dto.pot.SearchCondition.SearchRange;
 import org.com.itpple.spot.server.entity.Category;
 import org.com.itpple.spot.server.entity.Pot;
 import org.com.itpple.spot.server.entity.User;
